v3 overflow

+++ ft_memcpy.c
@@ -14,12 +14,12 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	if (!dest && !src)
 		return (NULL);
-	while (i < (int)n)
+	while (i < n)
 	{
 		((char *)dest)[i] = ((char *)src)[i];
 		i++;
